Doc openapi pour RentalController - Suppression du gestionnaire d'exception global pour remplacement par gestionnaire local - Price et Surface passent en Integer. Fix docopenapi UserController pour gestionnaire d'exception

+++ src/main/java/com/cpierres/p3backspring/entities/Rental.java
@@ -2,10 +2,6 @@ package com.cpierres.p3backspring.entities;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.CreatedBy;
-
-import java.math.BigDecimal;
-import java.util.Objects;
 
 //@Data//incompatible avec Auditable
 @EqualsAndHashCode(callSuper = true)
@@ -28,10 +24,10 @@ public class Rental extends Auditable {
     private String name;
 
     @Column(name = "surface", precision = 10)
-    private BigDecimal surface;
+    private Integer surface;
 
     @Column(name = "price", precision = 10)
-    private BigDecimal price;
+    private Integer price;
 
     @Column(name = "picture")
     private String picture;
